scsi: host: fix misaligned access
Creating a block device from a USB mass storage device fails with this
unliagned access

usb 2-3: New USB device found, idVendor=46f4, idProduct=0001, bcdDevice= 0.00
usb 2-3: New USB device strings: Mfr=1, Product=2, SerialNumber=3
usb 2-3: Product: QEMU USB HARDDRIVE
...
usb-storage 2-3:1.0: USB Mass Storage device detected
Oops - store (or AMO) address misaligned [#1]
...
[<ffffffff800713ba>] $x<capmode>.0+0x1a/0x6e
[<ffffffff806f135e>] usb_stor_probe1+0xa2/0x562
[<ffffffff806f2006>] storage_probe+0xae/0xf4

This points to the following code in usb_stor_probe1
   *pus = us = host_to_us(host);
   mutex_init(&(us->dev_mutex)),

The mutex_init causes the misaligned access. host_to_us returns the
hostdata component of a struct Scsi_Host. This returned a misaligned
struct in our case.

Make hostdata a properly aligned uintptr_t to fix the crash.

+++ include/scsi/scsi_host.h
@@ -727,8 +727,8 @@ struct Scsi_Host {
 	 * and also because some compilers (m68k) don't automatically force
 	 * alignment to a long boundary.
 	 */
-	unsigned long hostdata[]  /* Used for storage of host specific stuff */
-		__attribute__ ((aligned (sizeof(unsigned long))));
+	uintptr_t hostdata[]  /* Used for storage of host specific stuff */
+		__attribute__ ((aligned (sizeof(uintptr_t))));
 };
 
 #define		class_to_shost(d)	\
